+++ src/views/User/UserPolicy/index.tsx
